chore(storage): harden packed ownership boundary
